requirements.txt
streamlit
pillow
numpy
pandas
opencv-python-headless
tensorflow==2.13.0
matplotlib
# ipython